% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_funcs.R
\name{CG_plots}
\alias{CG_plots}
\title{Plot center of gravity of predicted distribution over the grid}
\usage{
CG_plots(
  data = test2,
  name = "step2",
  no_fish_zone = FALSE,
  folder = getwd(),
  conf,
  ...
)
}
\arguments{
\item{data}{data output from prepare_data function}

\item{name}{name to use for saving the output plots}

\item{folder}{defines save location}

\item{conf}{configuration file obtained from conf_TMB function}

\item{run}{model run output obtained from fit_mackerelST function}
}
\value{
CG plot
}
\description{
Plot center of gravity of predicted distribution over the grid
}
\details{
CG plot
}
