% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_funcs.R
\name{Marginal_plots}
\alias{Marginal_plots}
\title{Plot marginal effect of different variables}
\usage{
Marginal_plots(
  run,
  data = dat,
  conf,
  variable = "SST_0m",
  ref_year = 2018,
  folder = getwd(),
  ...
)
}
\arguments{
\item{run}{model run output obtained from fit_mackerelST function}

\item{data}{data output from prepare_data function}

\item{conf}{configuration file obtained from conf_TMB function}

\item{folder}{defines save location}

\item{name}{name to use for saving the output plots}
}
\value{
Plot indices of abundance
}
\description{
Plot marginal effect of different variables
}
\details{
Plot indices of abundance
}
