% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_funcs.R
\name{Plot_env}
\alias{Plot_env}
\title{Plot indices of abundance}
\usage{
Plot_env(data = test3, var = "SST_scl", truncate = FALSE, conf)
}
\arguments{
\item{data}{data output from prepare_data function}

\item{conf}{configuration file obtained from conf_TMB function}

\item{run}{model run output obtained from fit_mackerelST function}

\item{name}{name to use for saving the output plots}

\item{folder}{defines save location}
}
\value{
Plot indices of abundance
}
\description{
Plot indices of abundance
}
\details{
Plot indices of abundance
}
