% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_funcs.R
\name{plot_IA}
\alias{plot_IA}
\title{Plot indices of abundance}
\usage{
plot_IA(run, conf, IA_assess, name = "step2", folder = getwd(), ...)
}
\arguments{
\item{run}{model run output obtained from fit_mackerelST}

\item{conf}{configuration file obtained from conf_TMB}

\item{IA_assess}{index of abundance data.frame with the following column names ("Age", "Year", "IA_IESSNS")}

\item{name}{name to use for saving the output plots}

\item{folder}{defines save location}
}
\value{
Plot indices of abundance
}
\description{
Plot indices of abundance
}
\details{
Plot indices of abundance
}
\examples{
\dontrun{'blabla'}
}
