% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_funcs.R
\name{Plot_maps}
\alias{Plot_maps}
\title{Plot various map of predicted density, epsilon, etc on prediction grid}
\usage{
Plot_maps(
  pred_mu = test2,
  pred_epsilon = qwe2,
  name = "step2",
  no_fish_zone = FALSE,
  plot_data = FALSE,
  folder = getwd(),
  conf,
  ...
)
}
\arguments{
\item{name}{name to use for saving the output plots}

\item{folder}{defines save location}

\item{conf}{configuration file obtained from conf_TMB function}

\item{run}{model run output obtained from fit_mackerelST function}

\item{data}{data output from prepare_data function}
}
\value{
Plots various maps of outputs
}
\description{
Plot various map of predicted density, epsilon, etc on prediction grid
}
\details{
Plots various maps of outputs
}
