% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_funcs.R
\name{Plot_residuals}
\alias{Plot_residuals}
\title{Model diagnostics: QQplot of residuals}
\usage{
Plot_residuals(run, data, conf, name = "step2", folder = getwd(), ...)
}
\arguments{
\item{run}{model run output obtained from fit_mackerelST function}

\item{data}{data output from prepare_data function}

\item{conf}{configuration file obtained from conf_TMB function}

\item{name}{name to use for saving the output plots}

\item{folder}{defines save location}
}
\value{
residual qqplot
}
\description{
Model diagnostics: QQplot of residuals
}
\details{
residual qqplot
}
